find7:Add support for unified storage layout. Since Chinese ColorOS 2.1 use it.

+++ lvm/lvm_init.sh
@@ -11,8 +11,12 @@ if [ -e /dev/lvpool/userdata ]; then
     /sbin/static/busybox cp /fstab.qcom.lvm /fstab.qcom
     /sbin/static/busybox cp /etc/twrp.fstab.lvm /etc/twrp.fstab
     /sbin/static/busybox cp /init.fs.rc.lvm /init.fs.rc
-else
+elif [ -e /dev/block/platform/msm_sdcc.1/by-name/sdcard ]; then
     /sbin/static/busybox cp /fstab.qcom.std /fstab.qcom
     /sbin/static/busybox cp /etc/twrp.fstab.std /etc/twrp.fstab
     /sbin/static/busybox cp /init.fs.rc.std /init.fs.rc
+else
+    /sbin/static/busybox cp /fstab.qcom.unified /fstab.qcom
+    /sbin/static/busybox cp /etc/twrp.fstab.unified /etc/twrp.fstab
+    /sbin/static/busybox cp /init.fs.rc.lvm /init.fs.rc
 fi
