+++ lvm/lvm_init_recovery.sh
@@ -14,8 +14,12 @@ if [ -e /dev/lvpool/userdata ]; then
     /sbin/static/busybox cp /fstab.qcom.lvm /fstab.qcom
     /sbin/static/busybox cp /etc/twrp.fstab.lvm /etc/recovery.fstab
     /sbin/static/busybox cp /init.fs.rc.lvm /init.fs.rc
-else
+elif [ -e /dev/block/platform/msm_sdcc.1/by-name/sdcard ]; then
     /sbin/static/busybox cp /fstab.qcom.std /fstab.qcom
     /sbin/static/busybox cp /etc/twrp.fstab.std /etc/recovery.fstab
     /sbin/static/busybox cp /init.fs.rc.std /init.fs.rc
+else
+    /sbin/static/busybox cp /fstab.qcom.unified /fstab.qcom
+    /sbin/static/busybox cp /etc/twrp.fstab.unified /etc/recovery.fstab
+    /sbin/static/busybox cp /init.fs.rc.lvm /init.fs.rc
 fi
