+++ lvm/lvm_symlinks.sh
@@ -12,10 +12,15 @@ if [ -e /dev/lvpool/userdata ]; then
     /sbin/static/busybox ln -s /storage/emulated/legacy /mnt/sdcard
     /sbin/static/busybox ln -s /storage/emulated/legacy /storage/sdcard0
     /sbin/static/busybox ln -s /mnt/shell/emulated/0 /storage/emulated/legacy
-else
+elif [ -e /dev/block/platform/msm_sdcc.1/by-name/sdcard ]; then
     /sbin/static/busybox mkdir -p /storage/sdcard0
     /sbin/static/busybox chown root:root /storage/sdcard0
     /sbin/static/busybox chmod 0775 /storage/sdcard0
     /sbin/static/busybox ln -s /storage/sdcard0 /sdcard
     /sbin/static/busybox ln -s /storage/sdcard0 /mnt/sdcard
+else
+    /sbin/static/busybox ln -s /storage/emulated/legacy /sdcard
+    /sbin/static/busybox ln -s /storage/emulated/legacy /mnt/sdcard
+    /sbin/static/busybox ln -s /storage/emulated/legacy /storage/sdcard0
+    /sbin/static/busybox ln -s /mnt/shell/emulated/0 /storage/emulated/legacy
 fi
